TINY-1582: moved all bookmark code into separate dir
