+++ src/core/main/ts/bookmark/BookmarkTypes.ts
