+++ src/core/main/ts/bookmark/CaretBookmark.ts
@@ -12,7 +12,7 @@ import NodeType from '../dom/NodeType';
 import DomUtils from '../api/dom/DOMUtils';
 import Fun from '../util/Fun';
 import Arr from '../util/Arr';
-import CaretPosition from './CaretPosition';
+import CaretPosition from '../caret/CaretPosition';
 
 /**
  * This module creates or resolves xpath like string representation of a CaretPositions.
