+++ src/core/main/ts/bookmark/ResolveBookmark.ts
@@ -10,16 +10,16 @@
 
 import { Option, Options } from '@ephox/katamari';
 import Env from '../api/Env';
-import * as CaretBookmark from '../caret/CaretBookmark';
+import * as CaretBookmark from './CaretBookmark';
 import CaretPosition from '../caret/CaretPosition';
-import NodeType from './NodeType';
+import NodeType from '../dom/NodeType';
 import Tools from '../api/util/Tools';
 import { Selection } from '../api/dom/Selection';
 import { getParentCaretContainer } from 'tinymce/core/fmt/FormatContainer';
 import Zwsp from 'tinymce/core/text/Zwsp';
 import { DOMUtils } from 'tinymce/core/api/dom/DOMUtils';
 import CaretFinder from 'tinymce/core/caret/CaretFinder';
-import { isPathBookmark, isStringPathBookmark, isIdBookmark, isIndexBookmark, isRangeBookmark, PathBookmark, IdBookmark, Bookmark, IndexBookmark } from 'tinymce/core/dom/BookmarkTypes';
+import { isPathBookmark, isStringPathBookmark, isIdBookmark, isIndexBookmark, isRangeBookmark, PathBookmark, IdBookmark, Bookmark, IndexBookmark } from './BookmarkTypes';
 
 const addBogus = (dom: DOMUtils, node: HTMLElement) => {
   // Adds a bogus BR element for empty block elements
