+++ src/core/test/ts/browser/bookmark/CaretBookmarkTest.ts
@@ -1,6 +1,6 @@
 import { LegacyUnit } from '@ephox/mcagar';
 import { Pipeline } from '@ephox/agar';
-import * as CaretBookmark from 'tinymce/core/caret/CaretBookmark';
+import * as CaretBookmark from 'tinymce/core/bookmark/CaretBookmark';
 import CaretPosition from 'tinymce/core/caret/CaretPosition';
 import CaretAsserts from '../../module/test/CaretAsserts';
 import ViewBlock from '../../module/test/ViewBlock';
